backend/requirements.txt

Flask
Flask-JWT-Extended
python-dotenv
openai
PyPDF2
